requirements.txt
python-telegram-bot==21.6
pillow~=11.0.0
python-dotenv~=1.0.1
urllib3==2.2.3
